Work more seamlessly with 'set spell'
I found a nasty hack that allows us to have spellsitter work with
`set spell`. Basically we just need to define a syntax group called
Spell. As long as this is defined, the internal spellchecker will not
spellcheck the buffer.

This is much better than having `set nospell` which required us to apply
mappings to temporarily run `set spell` when running spell related
mappings.